refactor(playwright): UI signin helper を fixtures/ui_auth.ts に集約 + profile route spec 追加
helper 抽出:
- fixtures/ui_auth.ts に uiSigninAsRoot + RootFixture 型を export。viewport
  固定 → goto / → signin button click → username/password fill → submit →
  hydration check の 6 step を 1 関数に集約。
- post_note.spec.ts と authenticated_routes.spec.ts から duplicate を削除し、
  uiSigninAsRoot を import で利用。
- signin.spec.ts は signin flow 自体の smoke として inline duplicate を保持
  (= helper code path をテストする canonical spec)。

新 spec: authenticated_routes.spec.ts に /@alice navigation を追加。username
が body に visible で profile page が render されることを確認。

ローカル smoke + UI 計 20 件 pass を確認 (= 13 既存 + 4 auth route + 1 profile
+ 2 post_note/signin、合計 19.4s)。
